backend/requirements.txt
asgiref==3.8.1
azure-storage-blob==12.20.0
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
mysqlclient==2.2.4
PyJWT==2.8.0
ruff==0.4.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.6.0
